fix: prevent unwanted node selection by cleaning the recoil state used to trigger node selection after actually selecting a node
