+++ packages/twenty-front/src/modules/workflow/components/WorkflowDiagramCanvasEditableEffect.tsx
@@ -1,32 +1,19 @@
 import { useRightDrawer } from '@/ui/layout/right-drawer/hooks/useRightDrawer';
 import { RightDrawerPages } from '@/ui/layout/right-drawer/types/RightDrawerPages';
 import { useStartNodeCreation } from '@/workflow/hooks/useStartNodeCreation';
-import { workflowDiagramTriggerNodeSelectionState } from '@/workflow/states/workflowDiagramTriggerNodeSelectionState';
+import { useTriggerNodeSelection } from '@/workflow/hooks/useTriggerNodeSelection';
 import { workflowSelectedNodeState } from '@/workflow/states/workflowSelectedNodeState';
-import {
-  WorkflowDiagramEdge,
-  WorkflowDiagramNode,
-} from '@/workflow/types/WorkflowDiagram';
-import {
-  OnSelectionChangeParams,
-  useOnSelectionChange,
-  useReactFlow,
-} from '@xyflow/react';
-import { useCallback, useEffect } from 'react';
-import { useRecoilValue, useSetRecoilState } from 'recoil';
+import { WorkflowDiagramNode } from '@/workflow/types/WorkflowDiagram';
+import { OnSelectionChangeParams, useOnSelectionChange } from '@xyflow/react';
+import { useCallback } from 'react';
+import { useSetRecoilState } from 'recoil';
 import { isDefined } from 'twenty-ui';
 
 export const WorkflowDiagramCanvasEditableEffect = () => {
-  const reactflow = useReactFlow<WorkflowDiagramNode, WorkflowDiagramEdge>();
-
   const { startNodeCreation } = useStartNodeCreation();
 
   const { openRightDrawer, closeRightDrawer } = useRightDrawer();
   const setWorkflowSelectedNode = useSetRecoilState(workflowSelectedNodeState);
-
-  const workflowDiagramTriggerNodeSelection = useRecoilValue(
-    workflowDiagramTriggerNodeSelectionState,
-  );
 
   const handleSelectionChange = useCallback(
     ({ nodes }: OnSelectionChangeParams) => {
@@ -65,15 +52,7 @@ export const WorkflowDiagramCanvasEditableEffect = () => {
     onChange: handleSelectionChange,
   });
 
-  useEffect(() => {
-    if (!isDefined(workflowDiagramTriggerNodeSelection)) {
-      return;
-    }
-
-    reactflow.updateNode(workflowDiagramTriggerNodeSelection, {
-      selected: true,
-    });
-  }, [reactflow, workflowDiagramTriggerNodeSelection]);
+  useTriggerNodeSelection();
 
   return null;
 };
